requirements.txt
customtkinter>=5.2.2
keyboard>=0.13.5
pillow>=10.0.0
pystray>=0.19.5
